requirements.txt
flask
python-dotenv
requests
mysql-connector-python